% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis.R
\name{add_tf_hub_stats}
\alias{add_tf_hub_stats}
\title{Add Hub-Gene Counts to TF Statistics}
\usage{
add_tf_hub_stats(existing_stats, hub_genes)
}
\arguments{
\item{existing_stats}{A TF statistics data frame.}

\item{hub_genes}{A data frame containing hub genes and their module labels.
The name of the columns should be "gene" and "module".}
}
\value{
A TF statistics data frame with an added n_hub_genes column when
possible.
}
\description{
Adds, for each TF, the number of associated genes that are hub
genes.
}
