% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis.R
\name{clean_tf_names}
\alias{clean_tf_names}
\title{Normalize Transcription Factor Names}
\usage{
clean_tf_names(tf)
}
\arguments{
\item{tf}{A character vector with transcription-factor names.}
}
\value{
A character vector with normalized names.
}
\description{
Normalizes transcription-factor names by replacing common Greek
words to the alphabet lettersand removing dashes.
}
\keyword{internal}
