% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_clustering.R
\name{community_detection_master}
\alias{community_detection_master}
\title{Community detection wrapper}
\usage{
community_detection_master(
  adj_object,
  resolutions,
  number_iterations,
  seeds,
  graph_type = "snn",
  memory_log_file = NULL
)
}
\arguments{
\item{adj_object}{The adjacency matrix to be used. The matrix should be
square and should have defined rownames.}

\item{resolutions}{Check the \code{clustering_pipeline} documentation.}

\item{number_iterations}{Check the \code{clustering_pipeline} documentation.}

\item{seeds}{Check the \code{clustering_pipeline} documentation.}

\item{graph_type}{Check the \code{clustering_pipeline} documentation.}

\item{memory_log_file}{Check the \code{clustering_pipeline} documentation.}
}
\value{
A list of lists containing the clusters found for each combination
of number of neighbours and quality functions.
}
\description{
This function is a simplified version of the \code{clustering_pipeline}.
It starts with an already-generated adjacency matrix and applied the leiden
algorithm to identify the clusters. The details of how the algorithm is run
are the same as in the \code{clustering_pipeline} function.
}
