% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_processing.R
\name{compute_module_pairwise_tables}
\alias{compute_module_pairwise_tables}
\title{Compute Module Pairwise Tables}
\usage{
compute_module_pairwise_tables(module_mask, module_summ)
}
\arguments{
\item{module_mask}{A logical matrix of module masks.}

\item{module_summ}{A named list of module summary vectors.}
}
\value{
A list with pairwise overlap and correlation matrices.
}
\description{
Computes pairwise overlap, union and Spearman statistics between
the populations described by the module masks and summaries.
}
