% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monocle_plots.R
\name{diet_monocle_object}
\alias{diet_monocle_object}
\title{Diet Monocle object}
\usage{
diet_monocle_object(mon_obj)
}
\arguments{
\item{mon_obj}{Monocle object.}
}
\value{
Diet Monocle object.
}
\description{
This function simplifies the structure of the Monocle object.
The purpose is to create a lightweight object that can be used in the
context of the Shiny application. The resulting object will only contain
the needed information to perform the pseudotime ordering. THerefore,
this function removes information such as: the expression matrices, the
auxiliary data related to the dimensionality reduction and the clustering
and the metadata columns.
}
