% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{get_clusters_consistency}
\alias{get_clusters_consistency}
\title{Get the by-cluster consistency of configurations}
\usage{
get_clusters_consistency(by_cluster_list, order_logic = "avg_agreement")
}
\arguments{
\item{by_cluster_list}{A nested list of partitions that are grouped by
the number of clusters.}

\item{order_logic}{Determines how to order the partitions associated with
the same number of clusters. Can be either by frequency (most frequent first)
or by the agreement (the partition which is the most similar with the others,
first). Defaults to "agreement".}
}
\value{
A nested list with the partitions ordered by agreement / frequency
and the overall ECC value.
}
\description{
This function calculates the Element-Centric Consistency (ECC)
of the partitions that are grouped by the number of clusters.
}
