% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_matrix.R
\name{get_feature_loading}
\alias{get_feature_loading}
\title{Get feature loading from an expression matrix}
\usage{
get_feature_loading(expr_matrix, npcs = 30, approx = FALSE, ...)
}
\arguments{
\item{expr_matrix}{The gene by cell expression matrix to be used. It is
expected that the matrix is already normalized and scaled.}

\item{npcs}{The number of principal components to be used. Defaults to 30.}

\item{approx}{Logical indicating if deterministic PCA (with \code{prcomp}) or
approximate PCA (with \code{irlba}) should be used. Defaults to FALSE.}

\item{...}{Aditional parameters passed to the PCA function.}
}
\value{
A gene by npcs matrix containing the feature loading.
}
\description{
This function calculates the feature loading of an expression
matrix using PCA.
}
