% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_processing.R
\name{get_gene_overlap_stat}
\alias{get_gene_overlap_stat}
\title{Compute Gene Hub Statistics}
\usage{
get_gene_overlap_stat(
  expr_matrix,
  gene_modules,
  module_expr,
  gene_expression_threshold = 0,
  gene_expression_percentile = 0,
  module_expression_threshold = 0,
  module_expression_percentile = 0,
  scale = TRUE,
  total_weight = NULL
)
}
\arguments{
\item{expr_matrix}{A gene-by-cell expression matrix.}

\item{gene_modules}{A character vector of genes, or module-specific gene
list.}

\item{module_expr}{A numeric/logical vector, or a list of such vectors.}

\item{gene_expression_threshold}{Threshold for defining expressed genes.}

\item{gene_expression_percentile}{Optional percentile threshold for gene
expression.}

\item{module_expression_threshold}{Threshold for module activity.}

\item{module_expression_percentile}{Optional percentile threshold for module
activity.}

\item{scale}{Logical indicating whether values should be min-max scaled.}

\item{total_weight}{Optional vector of additional gene weights.}
}
\value{
A data frame with overlap metrics per gene.
}
\description{
Computes precision, recall and F1 overlap statistics between
the population induced by each gene and the overall population described
by the module it belongs to. If the gene weight is provided, the function
also calculates the combined score (defined as normalised weight * F1 score)
to determine hub genes.
}
