% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_processing.R
\name{get_per_module_weight}
\alias{get_per_module_weight}
\title{Compute Per-Module Gene Weights}
\usage{
get_per_module_weight(gene_adj_matrix, gene_modules)
}
\arguments{
\item{gene_adj_matrix}{A gene adjacency matrix.}

\item{gene_modules}{A named vector or list defining gene-module assignments.}
}
\value{
A named numeric vector of module-normalized weights per gene.
}
\description{
Computes normalized intra-module connectivity weights for
genes. For each gene, the sum of weights of the edges incident to it are
summed. These weights are normalised by the maximum weight of the module.
}
