% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{get_pseudotime_recommendation}
\alias{get_pseudotime_recommendation}
\title{Recommend a pseudotime ordering}
\usage{
get_pseudotime_recommendation(monocle_object)
}
\arguments{
\item{monocle_object}{A Monocle3 object.}
}
\value{
A list that contains the recommended metadata column, the subgroup
and the pseudotime values.
}
\description{
This function provides a recommendation of a pseudotime ordering
based on the metadata available in the monocle object. The recommendation is
done by selecting an end node from the diameter of the trajectory graph. The
function identifies the metadata group whose centre is the closest to the
starting point.
}
