% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{group_by_clusters_general}
\alias{group_by_clusters_general}
\title{Group the partitions by the number of clusters}
\usage{
group_by_clusters_general(clustering_list, start_level = 1)
}
\arguments{
\item{clustering_list}{A nested list of partitions. Each level is associated
with values of a specific parameters. The last level contains the partitions.}

\item{start_level}{The level at which the grouping should start. The default
is 1.}
}
\value{
A nested list of partitions grouped by the number of clusters.
}
\description{
This function groups the partitions by the number of clusters.
The grouping can be done at each level of the nested list by keeping the
hierarchy of parameters.
}
