% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{order_metadata_groups_by_pseudotime}
\alias{order_metadata_groups_by_pseudotime}
\title{Order Metadata Groups by Pseudotime}
\usage{
order_metadata_groups_by_pseudotime(metadata_value, pseudotime)
}
\arguments{
\item{metadata_value}{A vector with categorical metadata values.}

\item{pseudotime}{A numeric vector with pseudotime values.}
}
\value{
A character vector of metadata levels ordered by median pseudotime.
}
\description{
Orders categorical metadata levels by their median pseudotime
value.
}
