% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_clustering.R
\name{parallel_nn2_idx}
\alias{parallel_nn2_idx}
\title{Parallel NN indexing}
\usage{
parallel_nn2_idx(embedding, k)
}
\arguments{
\item{embedding}{The embedding matrix where each row represents a point
and each column a dimension.}

\item{k}{The number of nearest neighbours to be found.}
}
\value{
A matrix where each row contains the indices of the k-nearest
neighbours of the corresponding point.
}
\description{
This function provides a parallel approach of finding the
nearest neighbours of each points on a given embedding. The parallelisation
is done by splitting the data into chunks and running the \code{RANN::nn2}
function on each chunk in parallel.
}
