% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_matrix.R
\name{pca_reduction}
\alias{pca_reduction}
\title{Get PCA reduction from an expression matrix}
\usage{
pca_reduction(expr_matrix, npcs = 30, approx = FALSE, ...)
}
\arguments{
\item{expr_matrix}{The gene by cell expression matrix to be used. It is
expected that the matrix is already normalized and scaled.}

\item{npcs}{The number of principal components to be used. Defaults to 30.}

\item{approx}{Logical indicating if deterministic PCA (with \code{prcomp}) or
approximate PCA (with \code{irlba}) should be used. Defaults to FALSE.}

\item{...}{Aditional parameters passed to the PCA function.}
}
\value{
A gene by npcs matrix containing the PCA embedding.
}
\description{
This function calculates the PCA embedding of an expression
matrix.
}
\note{
The PCA is applied in this context to the genes, not to the cells.
}
