% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{plot_trajectory_graph}
\alias{plot_trajectory_graph}
\title{Plot a Trajectory Graph}
\usage{
plot_trajectory_graph(
  trajectory_object,
  edge_size = 0.5,
  edge_alpha = 1,
  edge_colour = "black",
  plot_nodes = c(1, 2, 3),
  node_size = 3,
  node_colours = stats::setNames(c("red", "blue", "green"), c(1, 2, 3)),
  node_label_size = 3,
  node_label_vjust = -1
)
}
\arguments{
\item{trajectory_object}{A trajectory object produced by
\code{get_trajectory_object}.}

\item{edge_size}{Line width for trajectory edges.}

\item{edge_alpha}{Transparency of trajectory edges.}

\item{edge_colour}{Colour of trajectory edges.}

\item{plot_nodes}{Which node-degree groups to display: 1 for leaves, 2 for
intermediate nodes, 3 for branching points. Setting to 0 will hide all nodes.}

\item{node_size}{Point size for trajectory nodes.}

\item{node_colours}{Named colours for node-degree groups.}

\item{node_label_size}{Text size of node labels.}

\item{node_label_vjust}{Vertical adjustment for node labels.}
}
\value{
A ggplot object of the trajectory.
}
\description{
Plots trajectory edges and optionally overlays selected node
degrees. This plot is meant to be used as an additional layer in an UMAP
plot.
}
