% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_umap}
\alias{plot_umap}
\title{Plot UMAP with Discrete or Continuous Coloring}
\usage{
plot_umap(
  umap_embedding,
  cell_info,
  mtd_name,
  cell_sort_order = c("lowest", "highest", "default"),
  scale_values = FALSE,
  cell_size = 0.3,
  cell_alpha = 0.8,
  legend_text_size = 10,
  legend_key_size = 0.5,
  axis_text_size = 10,
  show_labels = TRUE,
  label_size = 10,
  colourbar_width = 50,
  discrete_colors = NULL,
  continuous_colors = NULL
)
}
\arguments{
\item{umap_embedding}{A matrix or data frame containing the UMAP coordinates
of the cells. It should have two columns corresponding to UMAP1 and UMAP2.}

\item{cell_info}{A vector containing the information to color the cells by. It
can be either a factor/character vector for discrete coloring or a numeric vector
for continuous coloring. The length of this vector should match the number of rows in
\code{umap_embedding}.}

\item{mtd_name}{A string specifying the name of the metadata variable to be
used in the legend.}

\item{cell_sort_order}{A character vector specifying the order of cells based on
the \code{cell_info} variable. For discrete coloring, it should contain the unique values
of \code{cell_info} in the desired order. For continuous coloring, it can be either
"lowest", "highest" or "default" to sort cells by the \code{cell_info} values in
ascending, descending or no particular order, respectively.}

\item{scale_values}{A logical indicating whether to scale the \code{cell_info}
values between 0 and 1 for continuous coloring. Defaults to FALSE.}

\item{cell_size}{A numeric value specifying the size of the points in the plot.
Defaults to 0.3.}

\item{cell_alpha}{A numeric value between 0 and 1 specifying the transparency of
the points in the plot. Defaults to 0.8.}

\item{legend_text_size}{A numeric value specifying the size of the text in
the legend. Defaults to 10.}

\item{legend_key_size}{A numeric value specifying the size of the legend keys
for discrete coloring. Defaults to 0.5.}

\item{axis_text_size}{A numeric value specifying the size of the axis text in
the plot. Defaults to 10.}

\item{show_labels}{A logical indicating whether to show labels for discrete
groups. Defaults to TRUE.}

\item{label_size}{A numeric value specifying the size of the labels for
discrete groups. Defaults to 10.}

\item{colourbar_width}{A numeric value specifying the width of the colorbar
for continuous coloring. Defaults to 50 points. If set to 0, the colorbar
will be hidden.}

\item{discrete_colors}{A list of color vectors for discrete coloring, where the
names of the list correspond to the number of unique values in \code{cell_info}.
If NULL, the default ggplot2 colors will be used.}

\item{continuous_colors}{A vector of colors for continuous coloring. If NULL, the
default ggplot2 colors will be used.}
}
\value{
A ggplot object representing the UMAP plot with the specified
coloring and customization options.
}
\description{
This function plots a UMAP embedding with points colored by either
discrete or continuous cell information. The function provides options for
sorting the cells based on the coloring variable, adjusting point size and
alpha, and customizing the color scheme for both discrete and continuous
coloring.
}
