% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_utils.R
\name{read_gene_from_dense_h5}
\alias{read_gene_from_dense_h5}
\title{Read expression matrix from a dense HDF5 file}
\usage{
read_gene_from_dense_h5(
  gene_names,
  matrix_h5_path,
  index_genes = NULL,
  check_intersect = TRUE,
  add_rownames = FALSE
)
}
\arguments{
\item{gene_names}{List of names of genes to be read.}

\item{matrix_h5_path}{Path to the HDF5 file where the expression matrix
is stored. The file must contain the following fields:
\itemize{
\item genes: the names of the genes in the matrix.
\item cells: the names of the cells in the matrix.
\item expression_matrix: the expression matrix
}}

\item{index_genes}{Named vector where the names are the genes and the
values the index of the row associated with a gene. If NULL, the function
will calculate automatically this index. Defaults to NULL.}

\item{check_intersect}{Logical indicating if the function should check if
all the provided genes are present in the matrix. Defaults to TRUE.}

\item{add_rownames}{Logical indicating if the function should add the cell
names as column names for the extracted matrix. Defaults to FALSE.}
}
\value{
The expression matrix having the requested genes as rows.
}
\description{
This function read specific rows, associated with some genes,
from a dense matrix stored in a HDF5 file.
}
