% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_filtering.R
\name{remove_outlier_cells}
\alias{remove_outlier_cells}
\title{Remove outlier cells from a group}
\usage{
remove_outlier_cells(
  cell_names,
  umap_emb,
  percentile_threshold = 0.75,
  gmedian_point = NULL
)
}
\arguments{
\item{cell_names}{A vector with the names of the cells defining the group.}

\item{umap_emb}{A matrix with the UMAP embedding of the cells. The matrix
should have defined the rownames as the cell names.}

\item{percentile_threshold}{The percentile to be used as threshold for the
distance to the geometric median. This value will be used to calculate
the quantile. Defaults to 0.75.}

\item{gmedian_point}{The geometric median point of the group. If NULL, the
function will calculate this point. Defaults to NULL.}
}
\value{
A vector with the names of the cells that are not considered
outliers.
}
\description{
This function removes the outlier cells from a group based on
the distance patterns, as observed on an UMAP embedding. The function
considers a point to be outlier if the distance to the geometric median
of the group is above a given threshold determined by a quantile.
}
