% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_7_functional_assessment.R
\name{server_functional_assessment}
\alias{server_functional_assessment}
\title{Server - Functional Assessment}
\usage{
server_functional_assessment(id)
}
\arguments{
\item{id}{The id of the shiny module, used to access the UI elements.}
}
\description{
Creates the backend interface for the Functional Assessment panel
inside the Starlng Shiny application.
}
\note{
This function is a shiny module function and should be used
in the context of the app created using the \code{starlng_write_app} function.
}
