% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_gene_clustering.R
\name{server_gene_clustering}
\alias{server_gene_clustering}
\title{Server - Gene clustering}
\usage{
server_gene_clustering(id, filtered_genes)
}
\arguments{
\item{id}{The id of the shiny module, used to access the UI elements.}

\item{filtered_genes}{A reactive expression that contains the filtered genes
that will be used for the clustering.}
}
\description{
Creates the backend interface for the Gene Clustering panel
inside the Starlng Shiny application.
}
\note{
This function is a shiny module function and should be used
in the context of the app created using the \code{starlng_write_app} function.
}
