% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_gene_info_table.R
\name{server_gene_info_table}
\alias{server_gene_info_table}
\title{Server - Gene Info Table}
\usage{
server_gene_info_table(id)
}
\arguments{
\item{id}{The id of the shiny module, used to access the UI elements.}
}
\description{
Creates the backend interface for the Gene Info Tabel panel
inside the Starlng Shiny application.
}
\note{
This function is a shiny module function and should be used
in the context of the app created using the \code{starlng_write_app} function.
}
