% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_gene_heatmaps.R
\name{server_module_metadata_heatmap}
\alias{server_module_metadata_heatmap}
\title{Server - Gene Module Heatmap}
\usage{
server_module_metadata_heatmap(id)
}
\arguments{
\item{id}{The id of the shiny module, used to access the UI elements.}
}
\description{
Creates the backend interface for the Gene Module Heatmap panel
inside the Starlng Shiny application.
}
\note{
This function is a shiny module function and should be used
in the context of the app created using the \code{starlng_write_app} function.
}
