% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{starlng_write_app_clustassess_app}
\alias{starlng_write_app_clustassess_app}
\title{Create Starlng Shiny app from a ClustAssess object}
\usage{
starlng_write_app_clustassess_app(
  folder_path,
  ca_app_folder,
  stable_feature_type,
  stable_feature_set_size,
  stable_clustering_method,
  stable_n_clusters = NULL,
  use_all_genes = TRUE,
  app_title_name = "",
  learn_graph_parameters = list(nodes_per_log10_cells = 30, learn_graph_controls =
    list(eps = 1e-05, maxiter = 100)),
  gene_filtering_function = function(info_gene_df) {
     rownames(info_gene_df \%>\%
    dplyr::filter(.data$morans_I > 0.1, .data$q_value < 0.05))
 },
  clustering_parameters = list(n_neighbours = seq(from = 5, to = 50, by = 5), graph_type
    = "snn", prune_value = -1, resolutions = list(RBConfigurationVertexPartition =
    seq(from = 0.1, to = 2, by = 0.1), RBERVertexPartition = NULL,
    ModularityVertexPartition = NULL), number_iterations = 5, number_repetitions = 100),
  ecc_threshold = 0.9,
  freq_threshold = 30,
  save_entire_monocle = TRUE,
  discrete_colours = list(),
  continuous_colours = list(),
  max_n_colors = 40,
  verbose = FALSE,
  compression_level = 7,
  chunk_size = 100,
  nthreads = 1
)
}
\arguments{
\item{folder_path}{Check \code{starlng_write_app_monocle} documentation.}

\item{ca_app_folder}{The path of the folder containing the files associated
with the ClustAssess Shiny app.}

\item{stable_feature_type}{The feature type that is chosen from the
ClustAssess analysis.}

\item{stable_feature_set_size}{The size of the feature set that is chosen
from the ClustAssess analysis.}

\item{stable_clustering_method}{The clustering method that is chosen from the
ClustAssess analysis.}

\item{stable_n_clusters}{The number of clusters that is chosen from the
ClustAssess analysis. If NULL, the function will use all the clusters that
are obtained. Defaults to NULL.}

\item{use_all_genes}{If TRUE, uses all the genes in the expression matrix.
If FALSE, uses only the stable feature set as determined by ClustAssess.
Defaults to TRUE.}

\item{app_title_name}{Check \code{starlng_write_app_monocle} documentation.`}

\item{learn_graph_parameters}{Check \code{starlng_write_app_monocle} documentation.`}

\item{gene_filtering_function}{Check \code{starlng_write_app_monocle} documentation.`}

\item{clustering_parameters}{Check \code{starlng_write_app_monocle} documentation.`}

\item{ecc_threshold}{Check \code{starlng_write_app_monocle} documentation.`}

\item{freq_threshold}{Check \code{starlng_write_app_monocle} documentation.`}

\item{save_entire_monocle}{Check \code{starlng_write_app_monocle} documentation.`}

\item{discrete_colours}{Check \code{starlng_write_app_monocle} documentation.`}

\item{continuous_colours}{Check \code{starlng_write_app_monocle} documentation.`}

\item{max_n_colors}{Check \code{starlng_write_app_monocle} documentation.`}

\item{verbose}{Check \code{starlng_write_app_monocle} documentation.}

\item{compression_level}{Check \code{starlng_write_app_monocle} documentation.`}

\item{chunk_size}{Check \code{starlng_write_app_monocle} documentation.}

\item{nthreads}{Check \code{starlng_write_app_monocle} documentation.}

\item{expression_matrix}{A normalized gene by cell expression matrix. The
matrix should have the rownames and the colnames defined.}
}
\description{
Runs the entire Starlng pipeline on a normalized gene expression
matrix togheter with the content of the ClustAssess Shiny app
and generates a folder with all necessary files to run a Shiny app.
}
\note{
For more details about the other parameters and the content of the
shiny folder, consult the documentation of the \code{starlng_write_app_monocle}
function.
}
