% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{starlng_write_app_default}
\alias{starlng_write_app_default}
\title{Create Starlng Shiny app from a normalized gene expression matrix}
\usage{
starlng_write_app_default(
  folder_path,
  expression_matrix,
  metadata_df = NULL,
  pca_embedding = NULL,
  umap_embedding = NULL,
  app_title_name = "",
  learn_graph_parameters = list(nodes_per_log10_cells = 30, learn_graph_controls =
    list(eps = 1e-05, maxiter = 100)),
  gene_filtering_function = function(info_gene_df) {
     rownames(info_gene_df \%>\%
    dplyr::filter(.data$morans_I > 0.1, .data$q_value < 0.05))
 },
  clustering_parameters = list(n_neighbours = seq(from = 5, to = 50, by = 5), graph_type
    = "snn", prune_value = -1, resolutions = list(RBConfigurationVertexPartition =
    seq(from = 0.1, to = 2, by = 0.1), RBERVertexPartition = NULL,
    ModularityVertexPartition = NULL), number_iterations = 5, number_repetitions = 100),
  ecc_threshold = 0.9,
  freq_threshold = 30,
  save_entire_monocle = TRUE,
  discrete_colours = list(),
  continuous_colours = list(),
  max_n_colors = 40,
  verbose = FALSE,
  compression_level = 7,
  chunk_size = 100,
  nthreads = 1
)
}
\arguments{
\item{folder_path}{Check \code{starlng_write_app_monocle} documentation.}

\item{expression_matrix}{A normalized gene by cell expression matrix. The
matrix should have the rownames and the colnames defined.}

\item{metadata_df}{A data frame with cell metadata specific to the experiment.
If NULL, the function will create a one-group metadata with the name
"one_level".}

\item{pca_embedding}{A matrix with the PCA embedding of the cells. If NULL,
the function will calculate the PCA embedding while creating the monocle3
object.}

\item{umap_embedding}{A matrix with the UMAP embedding of the cells. If NULL,
the function will calculate the UMAP embedding while creating the monocle3.}

\item{app_title_name}{Check \code{starlng_write_app_monocle} documentation.}

\item{learn_graph_parameters}{Check \code{starlng_write_app_monocle} documentation.}

\item{gene_filtering_function}{Check \code{starlng_write_app_monocle} documentation.}

\item{clustering_parameters}{Check \code{starlng_write_app_monocle} documentation.}

\item{ecc_threshold}{Check \code{starlng_write_app_monocle} documentation.}

\item{freq_threshold}{Check \code{starlng_write_app_monocle} documentation.}

\item{save_entire_monocle}{Check \code{starlng_write_app_monocle} documentation.}

\item{discrete_colours}{Check \code{starlng_write_app_monocle} documentation.}

\item{continuous_colours}{Check \code{starlng_write_app_monocle} documentation.}

\item{max_n_colors}{Check \code{starlng_write_app_monocle} documentation.}

\item{verbose}{Check \code{starlng_write_app_monocle} documentation.}

\item{compression_level}{Check \code{starlng_write_app_monocle} documentation.}

\item{chunk_size}{Check \code{starlng_write_app_monocle} documentation.}

\item{nthreads}{Check \code{starlng_write_app_monocle} documentation.}
}
\description{
Runs the entire Starlng pipeline on a normalized gene expression
matrix and generates a folder with all necessary files to run a Shiny app.
}
\note{
For more details about the other parameters and the content of the
shiny folder, consult the documentation of the \code{starlng_write_app_monocle}
function.
}
