% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{subset_monocle_by_trajectory}
\alias{subset_monocle_by_trajectory}
\title{Subset the Monocle object by a trajectory}
\usage{
subset_monocle_by_trajectory(mon_obj, start_cells = NULL, end_cells = NULL)
}
\arguments{
\item{mon_obj}{A monocle object.}

\item{start_cells}{A list of cells names that define the start of the
selected trajectory. If NULL, no subsetting is performed. Defaults to NULL.}

\item{end_cells}{A list of cells names that define the end of the
selected trajectory. If NULL, no subsetting is performed. Defaults to NULL.}
}
\value{
The Monocle object subsetted by the trajectory.
}
\description{
This function subsets the Monocle object based on a subgraph
of the inferrred trajectory. The subgraph is defined using start and end
points, defined by cells.
}
