% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_gene_clustering.R
\name{ui_gene_clustering}
\alias{ui_gene_clustering}
\title{UI - Gene clustering}
\usage{
ui_gene_clustering(id)
}
\arguments{
\item{id}{The id of the shiny module, used to access the UI elements.}
}
\description{
Creates the UI interface for the Gene Clustering panel
inside the Starlng Shiny application.
}
\note{
This function is a shiny module function and should be used
in the context of the app created using the \code{starlng_write_app} function.
}
