% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_session.R
\name{ui_global_setttings}
\alias{ui_global_setttings}
\title{UI - Global Settings}
\usage{
ui_global_setttings()
}
\description{
Creates the UI elements with the global settings for the UI
components inside Starlng Shiny application.
}
\note{
This function is a shiny module function and should be used
in the context of the app created using the \code{starlng_write_app} function.
}
