% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_metadata_panel.R
\name{ui_metadata_umap}
\alias{ui_metadata_umap}
\title{UI - Metadata UMAP}
\usage{
ui_metadata_umap(id)
}
\arguments{
\item{id}{The id of the shiny module, used to access the UI elements.}
}
\description{
Creates the UI interface for the Metadata UMAP panel
inside the Starlng Shiny application.
}
\note{
This function is a shiny module function and should be used
in the context of the app created using the \code{starlng_write_app} function.
}
