% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_module_enrichment.R
\name{ui_module_enrichment}
\alias{ui_module_enrichment}
\title{UI - Module Enrichment}
\usage{
ui_module_enrichment(id)
}
\arguments{
\item{id}{The id of the shiny module, used to access the UI elements.}
}
\description{
Creates the UI interface for the Module Enrichment panel
inside the Starlng Shiny application.
}
\note{
This function is a shiny module function and should be used
in the context of the app created using the \code{starlng_write_app} function.
}
