% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_gene_modules_panel.R
\name{ui_module_umap}
\alias{ui_module_umap}
\title{UI - Gene Module UMAP}
\usage{
ui_module_umap(id)
}
\arguments{
\item{id}{The id of the shiny module, used to access the UI elements.}
}
\description{
Creates the UI interface for the Gene Module UMAP panel
inside the Starlng Shiny application.
}
\note{
This function is a shiny module function and should be used
in the context of the app created using the \code{starlng_write_app} function.
}
