% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_6_pseudotime.R
\name{ui_pseudotime_select_cells}
\alias{ui_pseudotime_select_cells}
\title{UI - Pseudotime}
\usage{
ui_pseudotime_select_cells(id)
}
\arguments{
\item{id}{The id of the shiny module, used to access the UI elements.}
}
\description{
Creates the UI interface for the Pseudotime panel
inside the Starlng Shiny application.
}
\note{
This function is a shiny module function and should be used
in the context of the app created using the \code{starlng_write_app} function.
}
