% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_session.R
\name{update_gears_width}
\alias{update_gears_width}
\title{Server - Gear Width}
\usage{
update_gears_width(session)
}
\arguments{
\item{session}{The shiny session object used to update the UI elements.}
}
\description{
Creates the backend interface that dynamically updates the
width of the gear settings window inside the Starlng Shiny application.
}
\note{
This function is a shiny module function and should be used
in the context of the app created using the \code{starlng_write_app} function.
}
