% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{update_monocle_partition}
\alias{update_monocle_partition}
\title{Update the partition of the Monocle object}
\usage{
update_monocle_partition(mon_obj, new_partition)
}
\arguments{
\item{mon_obj}{The Monocle object.}

\item{new_partition}{The new partition to be updated.}
}
\value{
The Monocle object with the updated partition.
}
\description{
This function updates the partition of the Monocle object,
which is used for tasks such as trajectory analysis.
}
