% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_0_session.R
\name{update_tabs}
\alias{update_tabs}
\title{Server - Tabs update}
\usage{
update_tabs(session)
}
\arguments{
\item{session}{The shiny session object used to update the UI elements.}
}
\description{
Creates the backend interface that dynamically shows the gene
module tabs inside the Starlng Shiny application after running the gene
clustering.
}
\note{
This function is a shiny module function and should be used
in the context of the app created using the \code{starlng_write_app} function.
}
