% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_utils.R
\name{write_gene_matrix_dense_h5}
\alias{write_gene_matrix_dense_h5}
\title{Write expression matrix to a dense HDF5 file}
\usage{
write_gene_matrix_dense_h5(
  expression_matrix,
  file_path,
  compression_level = 7,
  chunk_size = 100,
  all_at_once = TRUE,
  garbage_thresh = 2
)
}
\arguments{
\item{expression_matrix}{The expression matrix to be written. The matrix
should have defined rownames and colnames. If they are not defined, they
will be generated as \verb{gene_\{index\}} and \verb{cell_\{index\}} respectively.}

\item{file_path}{The path to the HDF5 file where the matrix will be stored.}

\item{compression_level}{The compression level to be used. Defaults to 7.}

\item{chunk_size}{The size of the chunks to be used. Defaults to 100.}

\item{all_at_once}{Logical indicating if the writing process will be done
by providing the entire matrix or by writing chunks of the matrix. It is
recommended to write the entire matrix at once, as no noticeable speedup
is observed in the other case. Defaults to TRUE.}

\item{garbage_thresh}{The threshold, measured GB, to trigger a garbage
collection while writing the matrix by chunks.}
}
\value{
The function does not return anything. It will create the HDF5 file
with the matrix stored in it.
}
\description{
This function writes a expression matrix to a HDF5 file in a
dense format. The created file will contain three fields:
\itemize{
\item genes: the names of the genes in the matrix.
\item cells: the names of the cells in the matrix.
\item expression_matrix: the expression matrix
}
}
