% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudotime.R
\name{get_pseudotime_recommendation}
\alias{get_pseudotime_recommendation}
\title{Recommend a pseudotime ordering}
\usage{
get_pseudotime_recommendation(
  monocle_object,
  recommendation_criteria = function(pseudotime_values) {
     fvn <-
    stats::fivenum(pseudotime_values)
     fvn[4] - fvn[2]
 }
)
}
\arguments{
\item{monocle_object}{A monocle object.}

\item{recommendation_criteria}{A function that defines the criteria of
determining the best pseudotime ordering. The function should take the
pseudotime values as input and return a numeric value. The function should
be monotonically increasing. Defaults to a function that returns the
interquartile range.}
}
\value{
A list that contains the recommended metadata column, the subgroup
and the pseudotime values.
}
\description{
This function provides a recommendation of a pseudotime ordering
based on the metadata available in the monocle object. The recommendation is
done by selecting the subgroup of a metadata column that leads to the highest
interquartile range of the pseudotime values (or any criteria the user
provides). This is translated as an ordering that has the highest
variability.
}
